requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
gunicorn==22.0.0
supabase==2.5.3
